+++ src/main/java/com/braintreegateway/Transaction.java
@@ -48,6 +48,7 @@ public class Transaction {
         CVV("cvv"),
         DUPLICATE("duplicate"),
         FRAUD("fraud"),
+        RISK_THRESHOLD("risk_threshold"),
         THREE_D_SECURE("three_d_secure"),
         TOKEN_ISSUANCE("token_issuance"),
         UNRECOGNIZED("unrecognized");
@@ -155,6 +156,7 @@ public class Transaction {
     private LocalPaymentDetails localPaymentDetails;
     private CustomActionsPaymentMethodDetails customActionsPaymentMethodDetails;
     private String planId;
+    private boolean processedWithNetworkToken;
     private String processorAuthorizationCode;
     private String processorResponseCode;
     private String processorResponseText;
@@ -283,6 +285,7 @@ public class Transaction {
             customActionsPaymentMethodDetails = new CustomActionsPaymentMethodDetails(customActionsPaymentMethodNode);
         }
         planId = node.findString("plan-id");
+        processedWithNetworkToken = node.findBoolean("processed-with-network-token");
         processorAuthorizationCode = node.findString("processor-authorization-code");
         processorResponseCode = node.findString("processor-response-code");
         processorResponseText = node.findString("processor-response-text");
@@ -530,6 +533,10 @@ public class Transaction {
 
     public String getPlanId() {
         return planId;
+    }
+
+    public boolean isProcessedWithNetworkToken() {
+        return processedWithNetworkToken;
     }
 
     public String getProcessorAuthorizationCode() {